extract ActiveTickets component
