+++ src/screens/Ticketing/Tickets/ActiveTickets/ActiveTickets.tsx
@@ -1,0 +1,67 @@
+import {StyleSheet} from '@atb/theme';
+import {
+  filterActiveOrCanBeUsedFareContracts,
+  isValidRightNowFareContract,
+  useTicketState,
+} from '@atb/tickets';
+import {TicketsTexts, useTranslation} from '@atb/translations';
+import useInterval from '@atb/utils/use-interval';
+import React, {useState} from 'react';
+import {View} from 'react-native';
+import TicketsScrollView from '../TicketsScrollView';
+
+export const ActiveTickets = () => {
+  const {
+    reservations,
+    fareContracts,
+    isRefreshingTickets,
+    refreshTickets,
+    didPaymentFail,
+  } = useTicketState();
+  const activeFareContracts = filterActiveOrCanBeUsedFareContracts(
+    fareContracts,
+  ).sort(function (a, b): number {
+    const isA = isValidRightNowFareContract(a);
+    const isB = isValidRightNowFareContract(b);
+
+    if (isA === isB) return 0;
+    if (isA) return -1;
+    return 1;
+  });
+
+  const hasAnyFareContractsOnAccount = fareContracts.length > 0;
+
+  const [now, setNow] = useState<number>(Date.now());
+  useInterval(() => setNow(Date.now()), 2500);
+
+  const styles = useStyles();
+  const {t} = useTranslation();
+  return (
+    <View style={styles.container}>
+      <TicketsScrollView
+        reservations={reservations}
+        fareContracts={activeFareContracts}
+        isRefreshingTickets={isRefreshingTickets}
+        refreshTickets={refreshTickets}
+        noTicketsLabel={t(
+          hasAnyFareContractsOnAccount
+            ? TicketsTexts.activeTicketsTab.noTicketsExpiredHelpText
+            : TicketsTexts.activeTicketsTab.noTickets,
+        )}
+        now={now}
+        didPaymentFail={didPaymentFail}
+        showTokenInfo={true}
+      />
+    </View>
+  );
+};
+
+const useStyles = StyleSheet.createThemeHook((theme) => ({
+  container: {
+    flex: 1,
+    backgroundColor: theme.colors.background_1.backgroundColor,
+  },
+  buyPeriodTicketButton: {
+    marginTop: theme.spacings.medium,
+  },
+}));
