+++ src/screens/Ticketing/Tickets/Tabs.tsx
@@ -1,25 +1,13 @@
 import {useAuthState} from '@atb/auth';
-import Button from '@atb/components/button';
 import {RootStackParamList} from '@atb/navigation';
 import {useRemoteConfig} from '@atb/RemoteConfigContext';
-import {StyleSheet, useTheme} from '@atb/theme';
-import {
-  filterActiveOrCanBeUsedFareContracts,
-  isValidRightNowFareContract,
-  useTicketState,
-} from '@atb/tickets';
-import {TicketsTexts, useTranslation} from '@atb/translations';
-import useInterval from '@atb/utils/use-interval';
 import {StackNavigationProp} from '@react-navigation/stack';
 import React, {useState} from 'react';
-import {TouchableOpacity, View} from 'react-native';
-import TicketsScrollView from './TicketsScrollView';
 import UpgradeSplash from './UpgradeSplash';
-import {AddTicket} from '@atb/assets/svg/mono-icons/ticketing';
 import {useAppState} from '@atb/AppContext';
-import {useHasEnabledMobileToken} from '@atb/mobile-token/MobileTokenContext';
 import {RecentTickets} from '@atb/screens/Ticketing/Tickets/RecentTickets/RecentTickets';
 import {AvailableTickets} from '@atb/screens/Ticketing/Tickets/AvailableTickets/AvailableTickets';
+import {ActiveTickets} from './ActiveTickets/ActiveTickets';
 
 export type TicketingScreenNavigationProp =
   StackNavigationProp<RootStackParamList>;
@@ -75,59 +63,3 @@ export const BuyTickets: React.FC<Props> = ({navigation}) => {
     </>
   );
 };
-
-export const ActiveTickets: React.FC<Props> = () => {
-  const {
-    reservations,
-    fareContracts,
-    isRefreshingTickets,
-    refreshTickets,
-    didPaymentFail,
-  } = useTicketState();
-  const activeFareContracts = filterActiveOrCanBeUsedFareContracts(
-    fareContracts,
-  ).sort(function (a, b): number {
-    const isA = isValidRightNowFareContract(a);
-    const isB = isValidRightNowFareContract(b);
-
-    if (isA === isB) return 0;
-    if (isA) return -1;
-    return 1;
-  });
-
-  const hasAnyFareContractsOnAccount = fareContracts.length > 0;
-
-  const [now, setNow] = useState<number>(Date.now());
-  useInterval(() => setNow(Date.now()), 2500);
-
-  const styles = useStyles();
-  const {t} = useTranslation();
-  return (
-    <View style={styles.container}>
-      <TicketsScrollView
-        reservations={reservations}
-        fareContracts={activeFareContracts}
-        isRefreshingTickets={isRefreshingTickets}
-        refreshTickets={refreshTickets}
-        noTicketsLabel={t(
-          hasAnyFareContractsOnAccount
-            ? TicketsTexts.activeTicketsTab.noTicketsExpiredHelpText
-            : TicketsTexts.activeTicketsTab.noTickets,
-        )}
-        now={now}
-        didPaymentFail={didPaymentFail}
-        showTokenInfo={true}
-      />
-    </View>
-  );
-};
-
-const useStyles = StyleSheet.createThemeHook((theme) => ({
-  container: {
-    flex: 1,
-    backgroundColor: theme.colors.background_1.backgroundColor,
-  },
-  buyPeriodTicketButton: {
-    marginTop: theme.spacings.medium,
-  },
-}));
